move component specific css to related css file
