+++ src/components/Modals/OrgSettingsModal/OrgSettingsModal.module.css
@@ -1,0 +1,5 @@
+.btns {
+    margin-top: 1rem;
+    display: flex;
+    flex-direction: column;
+}
