+++ src/components/Modals/PendingInvitationsModal/PendingInvitationsModal.module.css
@@ -1,0 +1,9 @@
+.invitationDisplay {
+    display: flex;
+    justify-content: space-evenly;
+    padding: 10px;
+}
+
+.inviteInfo {
+    margin-right: 20px;
+}
