+++ src/components/UI/CustomModal/CustomModal.module.css
@@ -40,28 +40,6 @@
 	width: 100%;
 }
 
-.invitationDisplay {
-    display: flex;
-    justify-content: space-evenly;
-    padding: 10px;
-}
-
-.usernameDisplay {
-    background: white;
-    color: black;
-    padding: 5px;
-    border-radius: 5px;
-    margin-right: 5px;
-    width: auto;
-}
-
-.usernameDisplayWrapper {
-    width: 100%;
-    display: flex;
-    flex-wrap: wrap;
-    margin-bottom: 15px;
-}
-
 .errorMsg {
     color: red;
     align-self: center;
@@ -91,30 +69,8 @@
     font-size: 1.25rem;
 }
 
-.descriptions {
-    font-size: .9rem;
-}
-
 .modalSubheader {
     font-size: 1.25rem;
     margin-top: 1rem;
     margin-bottom: 0;
-}
-
-.privateSection {
-    border: 1px solid white;
-    padding: 15px;
-    margin-top: 2rem;
-    border-radius: 5px;
-    margin-bottom: 1rem;
-}
-
-.inviteInfo {
-    margin-right: 20px;
-}
-
-.btns {
-    margin-top: 1rem;
-    display: flex;
-    flex-direction: column;
 }
